flask-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
anthropic==0.8.1
pandas==2.1.4
numpy==1.26.2
aiohttp==3.9.1
asyncio==3.4.3
geopy==2.4.1
python-Levenshtein==0.23.0
undetected-chromedriver==3.5.4
pydantic==2.5.2
httpx==0.25.2
tenacity==8.2.3